# optimize: consolidate pattern matching in security assessment (#312)